Added client to server communication
Note, the server's binary actuator with ID 2 seems to be problematic, if
you try to send a message to server with that ID the program will
(eventually) crash. No clue why. The server seems to also have a bug
regarding the Temperature actuator, if you try to change its' value
server will send a logout message stating the device is not an actuator.