Connectivity to multiple servers added
It should now be possible to connect to multiple servers, though for now
the code is changed in such a way that if you add additional servers,
they will be treated as "dummy" servers, they will not have any
networking capability and they will only show dummy data. This is to
showcase connectivity to multiple servers. Updated README to explain
these new things.